--- src/modules/completed.js
@@ -1,11 +1,0 @@
-import { finalTodo, storeItem } from './edit.js';
-
-export const checkedBox = (index) => {
-  finalTodo[index].completed = true;
-  storeItem();
-};
-
-export const notChecked = (index) => {
-  finalTodo[index].completed = false;
-  storeItem();
-};
